# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Local personal memory server: episodic (OpenSearch) + knowledge (Neo4j) derived from canonical hot JSON, archived to S3.",
        "title": "memory-mcp v2 API",
        "contact": {},
        "version": "2.0"
    },
    "host": "127.0.0.1:8420",
    "basePath": "/",
    "paths": {
...
        "/v1/consolidate": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ops"
                ],
                "summary": "Run the deterministic consolidation pipeline (§4)",
                "parameters": [
                    {
                        "description": "options",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ConsolidateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpserver.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/consolidate.Report"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpserver.Envelope"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/httpserver.Envelope"
                        }
                    }
                }
            }
        },
...
        "/v1/reindex": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ops"
                ],
                "summary": "Force full rehydration of derived stores (§5)",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "full hash audit",
                        "name": "verify",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpserver.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/rehydrate.Report"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/httpserver.Envelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apierr.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "consolidate.Candidate": {
            "type": "object",
            "properties": {
                "entities": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "episode_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "from": {
                    "type": "string"
                },
                "project": {
                    "description": "ws/team/proj",
                    "type": "string"
                },
                "to": {
                    "type": "string"
                }
            }
        },
        "consolidate.EntityStat": {
            "type": "object",
            "properties": {
                "co_occurrence": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "count": {
                    "type": "integer"
                },
                "entity": {
                    "type": "string"
                }
            }
        },
        "consolidate.Report": {
            "type": "object",
            "properties": {
                "archive_keys": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "candidates": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/consolidate.Candidate"
                    }
                },
                "entities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/consolidate.EntityStat"
                    }
                },
                "failures": {
                    "description": "Failures lists per-step errors; a failure never blocks other steps.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "moved_episodes": {
                    "description": "MovedEpisodes counts records shifted to cold; ArchiveKeys lists the\n{yyyy-mm}.json objects written.",
                    "type": "integer"
                },
                "snapshot_keys": {
                    "description": "SnapshotKeys lists knowledge latest+snapshot objects written.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "httpserver.ConsolidateRequest": {
            "type": "object",
            "properties": {
                "dry_run": {
                    "type": "boolean"
                },
                "projects": {
                    "description": "Projects limits the run to \"ws/team/proj\" strings; empty = all.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "httpserver.Envelope": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/apierr.Error"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "rehydrate.Report": {
            "type": "object",
            "properties": {
                "edges_upserted": {
                    "type": "integer"
                },
                "episodes_indexed": {
                    "description": "EpisodesIndexed / NodesUpserted / EdgesUpserted count replayed items.",
                    "type": "integer"
                },
                "failures": {
                    "description": "Failures lists per-project errors; partial success is reported, not\nhidden.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "nodes_upserted": {
                    "type": "integer"
                },
                "verified": {
                    "description": "Verified is set when verify=true ran a full hash audit (§5 reindex).",
                    "type": "boolean"
                }
            }
        }
    }
}